+++ src/routes/view/[year]/[month].svelte
@@ -84,6 +84,7 @@
 	export let monthData = {};
 	export let month, year;
 	$: pageHeader = `${getMonthString(month)} ${year}`;
+	$: currentBalance = monthData.totalIncome - monthData.totalExpenses;
 
 	/**
 	 * Sends an api call to delete an entry.
@@ -166,9 +167,9 @@
 <NewEntryForm on:new-entry={handleRefreshEntries} />
 
 <section>
-	<a sveltekit:prefetch class="flex items-center" href="/"
-		><ArrowLeft customClass="mr-1 inline" /> Back</a
-	>
+	<a sveltekit:prefetch class="flex items-center" href="/">
+		<ArrowLeft customClass="mr-1 inline" /> Back
+	</a>
 	<h1 class="text-2xl border-b-2 pb-2">{pageHeader}</h1>
 
 	{#await monthData}
@@ -185,6 +186,10 @@
 			<div>
 				<span class="font-black">Total Income: </span>
 				<span>{data.totalIncome ? currencyFormat(data.totalIncome) : '-'}</span>
+			</div>
+			<div>
+				<span class="font-black">Current Balance: </span>
+				<span>{data.totalIncome ? currencyFormat(currentBalance) : '-'}</span>
 			</div>
 		</div>